Factor out all that shiz into a hook
